Changes to state.py and adapting token.py to the changes
1. Make group_type the last parameter in the
beginGroup and endGroups calls, so that they can
take a default value. This is needed because the
BeginGroupToken and EndGroupToken classes do not
has access to the GROUP_TYPE class in state.py

2. Split the parameter array into two:
Parameters was renamed to LayoutParameters that
only store the layout parameters, and the other
control parameters do not need to track changes,
so they are stored in a separate dictionary.

3. Fixed a typo in the Domain class's __init__
method