Revert an earlier change for undefined commands
A previous change set a command to \relax when defining its meaning.
This is a problem because, in some cases, the command would be expanded
before it is properly defined, but after it is set to \relax. An example
would be \countdef\a=0\a=0. In this case, the second \a would be
expanded while reading the integer 0 (to find following digits and an
optional space), leading to a \relax token following the definition,
and thereby changing the meaning of the second \a.

The solution is to set the command to undefined (None), thus when the
second \a is expanded, it should return \a itself.

There is another problem left, that \edef\a{\b} will be successful,
which is not right. Here \b is not defined, so the \edef should fail.
This is now handled by checking the token when reading the balanced
text. If it is an undefined control sequence, we fail within
toks.readBalancedText.