Implemented \vcenter
To properly implement it, the import order has to be sorted out.
Originally, the box module imports hmode and vmode. However, this
cannot work, because eventually typesetting these lists need to
pack them into boxes, so they should depend on the box module. Yet
the box module depends on HList and VList to create the boxes. This
problem is solved by the asking the parser to create the lists.