Do not validate the command token
Validation was intended for prefixes for macrosm such as \long etc.
We validate by checking the next token. However, this is not correct
beause the prefix could be followed by another prefix, rending the
validation incorrect. Instead, we shoudl validate the value of the
assignment in the modify method. The exception produced by this shoudl
then be captured in the assign method, and the input stack location
will be attached to the exception.