Fixed a bug in conditional expansion
When reading a condition of an \if... command, the parser may meet
commands such as \else, \or or \fi. For example, in \if\odd1\else, the
\else command is expanded when scanner an integer. This is incorrect.
Thus, we make the branch statements not expandable during this stage.

To make it not expandable, we changed the way token_expand works.
Specifically, if the expand() method returns a token, then it is not
further expanded. So now, if we are parsing the condition, the branch
statements returns the current token.