Reimplemented the alignment handling
The original implementation read the tokens of the whole alignment
into a token list, then parse them. The problem is that, when parsing
each cell and read the tokens into a list, the readList function loses
the information about the context of being in an alignment cell. Thus
tokens such as &, \cr, and \span may not be handled correctly.

The new implementation creates an alignment builder, and puts it into
a stack, then resumes normal parsing. The commands such as \span, \cr,
and the & token are executed by the parser, which changes the state of
the alignment builder.

This also means that the scanners' terminate mechanism does not work
anymore with the right bracket token list scanner for a cell, because
it causes the input stack to return a None. Instead, now we use a stop
method, which defaults to None. When it is defined, it will be called
when the scanner reaches EOF. If it returns False, then the scanner
will be popped, and the parsing continues with the next scanner in the
input stack. Otherwise, the scanner will still be popped, but a None
token will be returned.

When a cell ends, its group is closed. The callback function for the
group calls the endCell method, which pushes the right bracket tokens
as a token list scanner with a stop method, onto the input stack, and
resumes parsing. When it finishes, the stop method calls the finishCell
method indicating that the cell is finished.